chore: created reusable sectioncard component
