+++ app/components/Views/TrendingView/TrendingTokensSection/TrendingTokensSection.tsx
@@ -1,30 +1,14 @@
-import React, { useCallback, useMemo } from 'react';
-import { View, StyleSheet } from 'react-native';
+import React, { useCallback } from 'react';
+import { View } from 'react-native';
 import { strings } from '../../../../../locales/i18n';
 import { TrendingAsset } from '@metamask/assets-controllers';
-import { useAppThemeFromContext } from '../../../../util/theme';
-import { Theme } from '../../../../util/theme/models';
 import TrendingTokensSkeleton from './TrendingTokenSkeleton/TrendingTokensSkeleton';
 import TrendingTokensList from './TrendingTokensList';
-import Card from '../../../../component-library/components/Cards/Card';
 import { useTrendingRequest } from '../../../UI/Assets/hooks/useTrendingRequest';
 import SectionHeader from '../components/SectionHeader/SectionHeader';
-
-const createStyles = (theme: Theme) =>
-  StyleSheet.create({
-    cardContainer: {
-      borderRadius: 12,
-      paddingVertical: 16,
-      paddingHorizontal: 16,
-      backgroundColor: theme.colors.background.muted,
-      borderColor: theme.colors.border.muted,
-    },
-  });
+import SectionCard from '../components/SectionCard/SectionCard';
 
 const TrendingTokensSection = () => {
-  const theme = useAppThemeFromContext();
-  const styles = useMemo(() => createStyles(theme), [theme]);
-
   const { results: trendingTokensResults, isLoading } = useTrendingRequest({});
   const trendingTokens = trendingTokensResults.slice(0, 3);
 
@@ -47,9 +31,9 @@ const TrendingTokensSection = () => {
           viewAllText={strings('trending.view_all')}
           onViewAll={handleViewAll}
         />
-        <Card style={styles.cardContainer} disabled>
+        <SectionCard>
           <TrendingTokensSkeleton count={3} />
-        </Card>
+        </SectionCard>
       </View>
     );
   }
@@ -61,12 +45,12 @@ const TrendingTokensSection = () => {
         viewAllText={strings('trending.view_all')}
         onViewAll={handleViewAll}
       />
-      <Card style={styles.cardContainer} disabled>
+      <SectionCard>
         <TrendingTokensList
           trendingTokens={trendingTokens}
           onTokenPress={handleTokenPress}
         />
-      </Card>
+      </SectionCard>
     </View>
   );
 };
